Conversion adds extra spaces before ampersands
This code is converted so that it has extra space before the ampersand `&`. That's because the regular expression is too liberal on matching and matches also zero characters.
```less
.example {
  color: red;
  &:hover {
    color: blue;
  }
}
```
Produces this result:
```scss
.example {
  color: red;
   &:hover {
    color: blue;
  }
}
```

With this fix from `*` to `+` I got better results and the ampersands were not misaligned. We use [stylelint](stylelint.io) to complain about these things and keep our code neat. That's how I spotted these and why it annoyed me. :)