Added support for converting LESS fade function
The LESS `fade(@mycolor, 50%)` function sets the opacity of an color to a specific value, percentage is supported. Our code had this in a few places. The matching code in SCSS is `rgba($color, 0.5)` and it does not support percentages. So I did a quick hack to convert the less code with percentages to `rgba($mycolor, 50%/100.0%)`. This way the conversion can be automated and it works in SCSS too.

I also added a basic match after that so that non percentage values do get converted too.